requirements.txt
# Core LangChain
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
langchain-google-community>=0.1.0

# Database and Vector Store
langchain-postgres>=0.0.6
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Email and Communication
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0

# Document Generation (for Drafting Agent)
python-pptx>=0.6.23
python-docx>=1.1.0
reportlab>=4.0.7
openpyxl>=3.1.2

# LangSmith (Monitoring and Tracing)
langsmith>=0.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type hints
typing-extensions>=4.9.0
